Created run_test.py
Script to replace secondary_warmer_script. This script has 3 test/ warmup modes that run sequentially.

Mode 1: sends an arbritrary (but ordered) list of trade signals through the API to test connection.

Mode 2: Pulls historical price data  of the 3 FX pairs from DB, uses it to backtest Algo. Trade signals are forwarded to the main API endpoint, trades and zones saved in DB.

Mode 3: Uses the zones from Mode 2 to "warmup" the live algo, such that it can start with a list of zones.